profile_validation/requirements.txt
asgiref==3.5.2
Django==4.0.5
djangorestframework==3.13.1
fuzzywuzzy==0.18.0
python-Levenshtein-wheels==0.13.2
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
